+++ bootique-cayenne/src/main/java/io/bootique/cayenne/DefaultDataSourceName.java
@@ -1,0 +1,19 @@
+package io.bootique.cayenne;
+
+/**
+ * A holder of an optional DefaultDataSource name.
+ *
+ * @since 0.18
+ */
+public class DefaultDataSourceName {
+
+    private String optionalName;
+
+    public DefaultDataSourceName(String optionalName) {
+        this.optionalName = optionalName;
+    }
+
+    public String getOptionalName() {
+        return optionalName;
+    }
+}