+++ RELEASE-NOTES.md
@@ -2,6 +2,7 @@
 
 * #15 Pure Bootique bootstrap of Cayenne projects
 * #23 Contributing Cayenne projects via DI
+* #26 Bad DataSource mapping results in an obscure Cayenne DI NPE
 
 ## 0.17